Implemented loop to process startup items
Filled in loop in function Submit-StartupItem to
process all the arguments for a startup item.

The loop first checks if there is an argument list
and then loops through the list. It uses a string
variable called $AllArgs to store the full list
and concatenates each argument to that variable.

The loop also checks if the startup item is a
browser and when on the last argument, which will
be the browser sites/tabs to open, adds them all
to the $AllArgs variable.

Finally, spaces are inserted so at the end of the
loop, $AllArgs is a properly formed string that
can be passed to Start-Process.