Filled out more of function Start-StartupItem and call it from Get-StartupItem
Set up proper parameter order for the function to
have $StartupItem (or $ItemNumber) first and
mandatory, $ProgramPath (or $ItemPath) second and
also mandatory, and $ArgumentsList (or $AllArgs)
as third.

I didn't make that last parameter mandatory and
gave it a default value of $null in case a startup
item doesn't have a arguments list.

Also updated Get-StartupItem to call
Start-StartupItem passing in the proper arguments.

Lastly, added if statement to Start-StartupItem
to check if $ArgumentsList has a value or not.
I will need to fill in the if blocks for each
scenario.