Added code to create startup.bat
Created variable to store the command to run the
startup.ps1 PowerShell script. Then directed that
variable to a file called test.bat. This is
currently being used to not clash with
startup.bat. However, in production, this will
change to redirect to startup.bat.

The idea is that when first "installing" this
program, CompStart, there will only be setup.bat
for less confusion. This batch file is run and it
will create startup.bat, move it to the user's
startup folder, and move all other program files
to a permanent directory (most likely in Program
Files).