Made it easier to switch between testing and production environments.
Removed the comment "**Uncomment when ready to put
into production**" (see commit
1b303cb898e8670816e50c4f3c7bde2fccb77d0a). Created
boolean variable called $IsProdEnv so only have to
change one location in code when switching between
production and testing use of startup.ps1 script.

Also added if blocks to the 2 locations where
previously there was the comment from commit
1b303cb898e8670816e50c4f3c7bde2fccb77d0a. Now,
depending on the value of $IsProdEnv, the main
do loop will prompt the user to run the script and
use the file startup_data.json, or assume the user
is always typing Yes and use the file
test_data.json.