Created schema and a text file
Using the json-schema.org specs, I created the
schema to validate startup_data.json. The schema
is currently not able to validate the key
"ArgumentsList" since it can contain both strings
and an array of strings.

Additionally, extra info such as required
properties, definitions, etc. need to be added.

Also create a misc.txt file to store the URLs of
two validation attempts.