Revert "Updated batch file name in setup.bat"
This reverts commit 1d7e7ecbae6b23c308734ed46befb3e4994285d7.

This commit, along with commits 09442 and 03a12,
are all to undo the changes made in commit 1d7e7.

In commit 1d7e7, the wrong variable was changed.
Instead of changing BATCHSCRIPT_FILENAME, the
variable POWERSHELLSCRIPT_FILENAME was changed.

In an attempt to revert the change, the commit
before it, commit 5e1942 was accidentally
reverted which produced commit 03a12. Then, in
order to revert commit 03a12, commit 09442 was
created.