Updated batch file name in setup.bat
Changed the BATCHSCRIPT_FILENAME variable from
test.bat, which was used in testing, to
startup.bat, which is being used in production.
Also updated the BATCHSCRIPT_RELPATH variable to
only "\CompStart\" since the
feature_addons/setup_startup no longer exists.

Ran the code and tested it worked. In the process,
it updated startup.bat to have the full path to
startup.ps1 for the "-File" parameter.