Set up feature branch features/bat_installer
Added content to the feature_addons/README.md for
this branch. Wrote up two paragraphs of what this
branch is for and the various files involved.

Also moved the setup.bat file from the project
root folder - CompStart - to the feature branch
folder - feature_addons/bat_installer.

Finally, updated setup.bat to the current path for
test.bat which is in data/test_data. Previously,
the path was still pointing to data/test_files
from a previous renaming of the test_data folder.