Started coding startupdata_editor.py
Added initial code to switch between testing and prod environments.
Created function json_reader to read in the startup_data.json file for
prod or the test_data.json file for testing. Added code to the function
to test to make sure the filename given to open is a JSON file (i.e., it
has a .json extension).

Need to refactor the if-block that checks the filename as it can be
simplified and made more concise.