Finished json_reader function in startupdata_modifier_tool.py
Finished the function json_reader so it correctly
either reads in the JSON data for a valid JSON
file, or it returns an error message.

Also added functionality to main that if an error
message is returned, print that message and exit
the program with a exit status of 1.

Finally, had to remove the copy of test_data.json
in the feature branch folder and use the original
copy in data/json_data because part of the code to
read the file involved using the os module along
with os.path.join, os.getcwd and os.sep.