Minor linting change and addition of main loop
Split the error message returned when the file
extension is not .json to use Python's bracketed
multi-line string syntax so the string didn't
break the 80 character line rule.

Added a loop in main to give the user a choice to
create a new JSON file, edit an existing one or
quit. Coded into the loop the ability to handle
invalid user choices. Also, set up call to
function json_reader to happen when user chooses
to edit an existing JSON file.

Lastly, added else to the line if not is_json_file
so that if the file is a valid one, a message
prints out.