Several changes - see details
Created a new function called json_creator to
call when the user selects option 1 from the main
loop. Similar to json_reader, it takes in the
json_path list and the json_filename string
variables.

Also created a helper function called
parse_full_path to assist with parsing json_path
and json_filename into one absolute path in string
format. Adjusted json_reader to call this helper
function.

Made some changes to json_reader including
renaming the is_json_file boolean variable to
read_json_success. This way, in the try-except
block while reading in the JSON data, if there's
an exception, a message could be returned "Unable
to read JSON data" and the boolean is only set to
true when the JSON data has been read in
successfully.

Finally, adjusted the call to the
helper function as well as the try-except block to
all be under the nested if-else block as part of
the else clause. This makes sure that only if the
JSON file has a .json extension, will reading the
data be attempted.