Created an Enum class to help with generating the necessary JSON startup data
Created an Enum class called JsonSchema which will
has as its members the valid keys that can be used
for the JSON startup data. Also created a global
(constant) variable called EXAMPLE_JSON which is
just the contents of test_data.json as a string.
This will hopefully make it easier to create the
code to generate JSON data using the JsonSchema
class since there is now an example to work with.