Changed EXAMPLE_JSON variable and wrote it back to a JSON file
Played around with the functions json.dump and
json.dumps to understand the differences between
writing JSON data as a string to a file versus not
as a string. Realized it's better to not make it a
string but use Python's data types of dictionary
and lists which the dump function will convert
automatically to a JSON object and array.

Changed EXAMPLE_JSON so instead of a string, it is
now a dictionary containing all the JSON data from
test_data.json. Also tested writing that back to a
file called temp.json to confirm the data is
written correctly.