Renamed JsonSchema enum class and created new enum class
Renamed JsonSchema to JsonSchemaKeys because that
enum class just contains the valid keys that can
be used for startup_data.json.

Create new enum class called JsonSchemaStructure
with two members: OBJECT and ARRAY. OBJECT is a
dictionary and ARRAY is a list. This can be used
to generate the structure of the startup_data.json
JSON data.