Created two new functions to help with creating the JSON data and updated the create_json_data function
Updated the create_json_data function docstring to
specify what the function does and the arguments
it takes. Created two helper function for actually
generating the JSON data, generate_default and
generate_json.

The first one will create the
default initial JSON startup data for a new
startup_data.json file. The second will be used
when editing an existing file to update it.

Filled out the docstring for both helper functions
and added logic to create_json_data to determine
which helper function to call.