Filled out all the code for the generate_default function
Used the two Enum classes and their members to
create a basic startup JSON object consisting of
1 item, to open notepad. Used
startup_data.schema.json as a guide and populated
the json_data dict.

Along the way, realized that because dicts and
lists are mutable, when adding them to json_data,
need to copy the members from the Enum class
JsonSchemaStructure using their values. For
example, JsonSchemaStructure.ARRAY.value.copy().