Fleshed out comments and minor updates to return statements
Added more to the generate_default function
docstring to list out what the default JSON
startup data actually is. Also added a docstring
comment above the code that populates json_data to
explain in detail what the code is doing.

Finally, added a return statement for the
generate_default function and two print statements
to the create_json_data function to confirm the
code is running correctly.