Major changes startup_data.schema.json and minor changes to extra_info.txt
Added to new validation attempts to extra_info.txt
based on changes made to startup_data.schema.json.

For the schema, added a title and description key
just under the $schema key. Added a default key to
the ArgumentList parameter as well as a required
key to both the object within the Items array and
the root object. The two required keys both say
that all parameters are required. Also added a
uniqueItems key to the Items array, so all objects
with that array will need to be unique.