Made updates to reflect commit f1aec23 on feature branch features/json_schema_update
On commit f1aec23 for feature branch
features/json_schema_update, a new property was
added to the startup_data.json schema called
"Name". It comes right after "ItemNumber" and
before "FilePath".

The Enum class JsonKeySchema was updated to
reflect that as well as the code in
startupdata_editor.py in all the places that
needed to be updated. Lastly, test_data.json
within this feature branch was updated to reflect
the new schema.