Made several changes to the function json_creator
Added in a return statement to return whether the
file was written successfully or not. Created two
variables at the beginning of the function and
initialized them with values. Also, updated the
function docstring.

Finally, added an explicit dict conversion to the
output of the create_json_data function call to
satisfy an error from Pylance and removed the
comment # type: ignore.