Refactored code for writing JSON data
In the json_writer function, moved the return
message that JSON file was written successfully
to here and made it the default message. Also,
added a return message when the file already
exists and the user chooses not to overwrite it.

In the json_creator function, renamed the
status_message variable to return_message to put
it in line with all the other functions that
return a message. Created a file_state variable
and refactored the if-block to only check if the
JSON file exists. If so, the file_state variable
is set to 1 but it has a default of 0. Also moved
called create_json_data to outside the if block.
Finally, moved calling json_writer to outside the
if block and pass in file_state instead of a hard
coded value. This ensures that json_writer is the
only place where check_overwrite is called when
file_state is 1.

In json_reader, added a default return message of
the data read being successful. Finally, in the
main loop, set it to print the return or status
message and also print the JSON data when
returning from json_reader and the read was
successful. This was done for testing purposes.