Refactored code in main loop
Created a new variable called status_state and use
that for the return statuses of the calls to
json_creator and json_reader. Took out the print
status_message statement from the if-block for
user's input and put it underneath that block.
Also added code to only print status_message when
the user hasn't selected to quit the program.