Refactored code for json_writer
Created new variable called file_mode and use that
in the match block setting it to "w" or "a". As a
result, took out the try-except block to write the
actual JSON data from the match block and put it
underneath on its own. Set the file_mode variable
to default of a blank string and use that for the
base case in the match block. See diff for
details.