Changed code in function json_creator to test if the JSON file exists
Was previously using a simple if json_file where
json_file was a string variable containing the
full absolute path to the JSON file. While
attempting to test the creation of new file by
renaming the existing test_data.json to
test_data.json.orig, discovered that the if
statement doesn't work.

Refactored that code to use the os.path.isfile
function to properly test if the file exists or
not. This passed the test of renaming
test_data.json to test_data.json.orig. Adding both
test_data.json and test_data.json.orig as well.