Several changes
Updated ideas.txt to enclose the idea(s) I've
implemented in square brackets with the word
"Done" in front, and the ones I still need to
add with curly brackets with the words "To Do" in
front.

Also, updated both startup_data.schema.json and
startup.ps1 to use "Description" instead of
"Comments" per a suggestion in ideas.txt.

As a result, there are only now 2 suggestions from
ideas.txt that need to be implemented - to add the
JSON schema key "description" to each JSON
parameter with a description of what each
parameter is or means; and to use the parameters
"Name" and "Description" in the Python program to
edit the JSON startup data.