Updated startup_data.schema.json
Added a few "default" schema keys as well as
updated the "examples" schema value to reflect
recent changes to the schema itself. The example
now contains a "Name" and "Description". Finally,
updated the values for one of the "required"
schema keys to make sure it also had "Name" and
"Description".