Renaming of variable and test function
Renamed the variable cls.TEST_DICT to
cls.EXAMPLE_TEST. The plan is to create a second
set of variables starting with cls.TEST equivalent
to the cls.JSON variables. These can be used to
help with testing.

Also renamed the 'test_parse_full_path' function
to 'fntest_parse_full_path'. For now, the
nomenclature will be to use 'untest' for the test
functions that I haven't gone through one by one
to confirm the test itself is properly working.
When I'm working on a test function, I will remove
the 'un' part, and when I'm done confirming the
test function works, I'll add 'fn'.