Adjusted class constant variables
Created cls.TEST constants equivalent to the
cls.JSON variables as mentioned in the previous
commit. Rearranged the set of cls.JSON variables
to be together and the set of cls.TEST variables
to be together.

Also changed cls.EXAMPLE_TEST to follow the
startup data JSON schema but added 2 items:

1. Calculator
2. Dealer-FX homepage in a new Chrome window