Added some class non-constant variables
Created the following variables:

cls.expected_message, cls.sde_func_str_return,
cls.sde_func_tpl_return, cls.sde_func_dict_return

The first variable will be for each test function
to use and print if the test fails. As the
variable name implies, it will store the expected
value for the test function.

The last three variables will be used by each test
function to store the return values of the call to
the startup_data_editor.py function that is being
tested. This way, the return values can be printed
if desired and it also makes for cleaner code in
the self.assertEqual function calls.