Made some test function name changes
Renamed the 3 test functions confirmed working
from commits 23f3a4e, f0dc796, and 6acd89d to
fntest as the beginning part of the function name
per details from commit 99a8b25.

Also decided to rename the test functions that
aren't yet filled in to notest for the beginning
part of the function name.

Finally, changed the function
test_create_json_data_false to notest and removed
all the code in it except a print statement
saying this test is being skipped. This is because
the function create_json_data from
startup_data_editor.py isn't implemented to take
a value of False when passed in.