Updated the test_json_writer_case_zero function
Rearranged the existing code to create more space
between the lines. Also added new code to first
check if the file that was supposed to have been
created actually exists. Next, the code reads back
in the JSON data from that file and compares it to
what should have been written to confirm. If
either check fails, an error message is printed.