Worked on function test_json_writer_case_two
Rewrote code for test_json_writer_case_two to
read back in the file after being appended to and
compare the data. Used self.TEST_FILE as the data
file and wrote self.EXAMPLE_JSON as the append
data.

Initially, kept running into errors when reading
back the data despite the test call to function
json_writer in startup_editor.py with a file_state
of 2 to indicate append returning True.

Tried to compensate by erasing the data file and
writing into it from scratch, similar to calling
test_json_writer_case_zero, and then proceeding.
Tried also debugging but got the same results.

Finally, ran an interactive Python shell for just
startup_editor.py and found out the issue is that
in json_writer, the attempt to call json.dump with
the data file open for append results in an error.
There's something about how the data is first
written that adds an EOF marker and when trying to
append, calling json.dump again just results in
the new data being written after the EOF marker.

However, when running the test function, none of
that error information from the json.dump function
is received.

Committing all this code to history for posterity,
but will need to rewrite the "append" aspect to
what it really should be: code that reads in the
current data from the JSON file and updates the
dictionary programmatically and then writes it
back. This is because even using the built-in
dictionary method update doesn't work since the
keys are the same. To see an example, run
from the test.py file the command
"EXAMPLE_TEST.update(EXAMPLE_JSON)". This will
cause all of the data in EXAMPLE_TEST to be
overwritten completely with the data in
EXAMPLE_JSON.