Updated fntest_json_creator and rewrote JSON data for unittest_data.json
Refactored fntest_json_creator slightly to use
self.TEST_FILE when confirming that the JSON data
was written successfully. This is because in the
first part of the test function, when calling
json_creator from startup_data_editor.py, the
self.TEST variables were passed in. This also
caused the default JSON data, found in
self.EXAMPLE_JSON, to be written to
unittest_data.json.