Adding onto commit 3085c47
Commit 3085c47 didn't include all the changes made
and mistakenly mentioned that the json_writer
function code changes still had to be made.
However, commit 3085c47 contains both the
docstring and code changes to json_writer.
Unfortunately, commit 3085c47 was already pushed
to GitHub, so this new commit is being used
instead. There are no actual changes being
committed.

So, in commit 3085c47, I refactored the code in
case 2 to now read in the JSON file and then check
to see if it's the same as what's been passed in.
If it isn't, we set the file_mode to "w" because
there's updated JSON data to write. This ensures
that when the case is 2, while a file read
operation is performed, a file write operation
isn't performed if not necessary.