Updated the docstring for json_writer
Added to the docstring for the function
json_writer about what the function will do when
the 'file_state' parameter passed through is '2'.

Previously, it was meant to append but that's not
possible anymore. Read the docstring to learn the
current way the function will operate. Need to
actually code in the changes.