Updated case 2 try-except block
While testing, noted that the variable
existing_data, which is used to store the existing
JSON file information when read in, may not be
bound due to it only being assigned anything in
the try block and also not being declared
elsewhere. Added a line in the except block to
set existing_data to be the same as the json_data
variable, which still allows the if statement
after the try-except block to work.