Resubmitting changes from commit 8e013ad along with the code change I forgot to include in that commit.
Renamed the parameter from 'default' to 'new_file'
which reflects better the purpose of the
parameter. This way, it's easier to tell that when
the parameter is True, then a new JSON file will
be created and if false, then the file already
exists. In total, 3 lines of code were changed.