Made test_check_overwrite_no as its own function
Copied test_check_overwrite_yes and changed mock
input to "N". Then ran all tests - 12 in total at
this point - and confirmed they all run
successfully as well as print out the relevant
output necessary, such as for the tests that
cause JSON data to be written to file, checking
that, after calling the function in
startup_data_editor.py, the data written is
correct in addition to the function returning as
expected.