Added new non-class function
Created helper function to set the start directory
to be the program CompStart root folder even if
new_test.py is run from a different working
directory. This is called from the main if
statement and isn't part of the Test class.

This function also assumes "CompStart" is
somewhere in the working path and doesn't error
check for that. This was done because new_test.py
will always be called from within a project
directory.