Figured out how to order the tests
After some trial and error, determined how to
order the tests to set in which order they run.
Adjusted the print statements at the beginning of
each function as a result. Also removed the
'notest' for the two tests not yet written so they
still run as part of the test suite but just
output that the function being tested is being
skipped.