Added some comments above each test function
Above each test function, added the type of
return_value variable needed or used. The plan
will be to see if one or more helper functions
can be created to assist since 2 of the test
functions use Boolean, 4 use Dictionaries, 7 use
Tuples and 1 uses String.

There are also other areas in some of the test
functions that have duplicated code that can be
farmed out to a helper function.