Created helper function confirm_written_data
Created helper function to read in the contents of
unittest_data.json and compare it to whatever
JSON data was supposed to be written to the file.
This is called by the test_json_creator function
and the 3 test_json_writer functions. They each
pass to this helper function the dictionary to
test against.