Created helper function print_results
Created helper function to print both the
"Expected: " and "Actual: " messages at the end of
each test function. Updated all the functions to
call this helper function, passing in the specific
data to print out for the "Actual: " line. The
helper function casts that data passed in to a
string.