Created new helper function
Created function clear_test_file for the tests
that write JSON data to file. These are the
test_json_creator and the 3 test_json_writer
functions. In all four of them, there was an if
statement checking for the existence of the file
and removing the file if it already existed. That
if block is now the helper function
clear_test_file.