Refactored code to have one function called test_suite that calls all the other functions
Renamed all the other "test" functions to start
with "fn" and removed the numbers. In the
"test_suite" function, called all the actual
"test" functions in the order they should be. Did
this so it will be easier to reorder functions,
create new ones or skip some when fixing a bug.

As part of the process, had to create a new helper
function to reset the variables:
self.expected_message /
self.sde_func_str_return /
self.sde_func_tpl_return /
self.sde_func_dict_return /
self.sde_func_bool_return

Updated all the "fn" functions to call this helper
function as the first thing so it wouldn't mess up
the above variables.

Finally, rearranged the placement of the helper
functions with respect to the code and also the
ordering of the "fn" functions.