Changed how the main loop is handled and created a new helper function
Changed the main loop logic from an if-elif-else
block to a match-case block. Also moved the print
statement of the variable status_message from
outside the conditional block to inside each case
block.

Finally, created a function called
print_pretty_json which is called when the user
chooses to view the existing JSON data. Currently,
this function just calls json.dumps and adds an
indent of 2. Will need to rewrite the code to
go through the JSON data dictionary and properly
parse it.