Created helper function prettify_error and other changes
Created a helper function to use for all
try-except blocks. This way, the except block can
call this function and this avoids duplication of
code.

Also updated the doc strings for the functions to
be more consistent in indenting and spacing. Some
docs strings were also updated to indicate the
function is a helper function.

Changed the main loop to not quit if the user
chooses 3 but just display that ability isn't
available. Also fixed error where the numbers
considered acceptable choices were hardcoded in.
Instead, created a variable max_choices to easily
update if the number of choices in the main loop
changes.