Made a change to the logic of the main loop to fix a bug
Realized that when getting the user input for the
main loop, if the user entered an non-numeric
character or word, an error was thrown because
the line of code casting the user input to an
integer needed to be in an else clause for the
if statement checking for valid input.

Fixed that and confirmed it now works. Any input
that's not a valid choice, which is currently 1-4,
shows the "Please enter a valid choice" output and
continues the loop.