Changed how the test functions are ordered and numbered
Created a class variable that acts as a counter
and is incremented every time generate_message is
called at the beginning of a test function. This
means the test functions don't need to pass in a
test number and can be called in any order but
the printout will still say "Test #" in order.