Randomized order of test functions and created ability for user to pause between each test
Created a loop that will go through all the test
functions but in random order every time. Also
added question to ask user if they want to pause
in between each test. The default is no, but if
user enters "Y" or "y", after each test function
is run, the user will be asked to type a key to
continue. This will allow them to read the output
of each test if so desired.