Tested displaying startup data with production startup_data.json file
Came across two use cases that caused issues. The
first was when the arguments list has only 2
arguments, so changed the if block in
prettify_startup_item to treat an argument count
of 2 the same as more than two. The second was
when the startup item is a browser and the last
argument passed in is a list of websites to open.
Added in an if statement to check for that and
use the string join function to convert the list
into a string with a space between each website.