Started work for case 3 of the main loop
Created function json_editor in similar fashion to
json_reader and json_creator. Started coding the
function and currently it will print out the
startup data and then ask the user which startup
item they want to edit. Will loop through until
valid input is given, which is a number
corresponding to an actual startup item or "Q" to
go back to the main menu.