Added comments, docstrings and two updates to json_editor
Created a docstring for json_editor as well as
added comments throughout the code. Also added the
option to type "A" if a user wants to add a
startup item. Finally, moved the code to display
the startup data and loop for user input to run
only if the call to json_reader was successful.