Added code to parse user input for json_editor loop
Filled out the if block for when the user enters
valid input. First the input is formatted. If the
user entered "q" or "a", convert it to uppercase.
If the user entered a startup item number, convert
that to an integer. Next, a match-case is used
with currently, case "Q" setting the loop to quit,
case "A" printing the functionality hasn't been
implemented and then setting the loop to quit, and
the default case being to print out the startup
item number selected.