Added new helper function called edit_startup_item
This helper function is called by the json_editor
main loop when a user selects to edit the startup
data. Commented out several lines and updated some
comments to make the flow of things a bit
different. Now, when a user selects to edit the
startup data from the initial main loop, the total
number of items is displayed along with the
choices added in the last commit. Once a user
selects a startup item number, then the details
of that startup item is displayed.