Made a change to how startup items are printed out and the loop for edit_startup_item
Added in prettify_startup_item a line to print out
the item program or file path as well as the other
data. It doesn't print out if the item is a
browser or not, but the plan will be to determine
that during the adding or editing of a startup
item.

Also changed the loop in edit_startup_item to be
more like the loop in json_editor in terms of the
logic.