Made changes to the JSON being tested
Changed the dictionaries storing JSON data set up
in the setUpClass function to be lists instead.
This will make it easier to update should the JSON
data schema changes.

Updated the test functions that compare file JSON
data to these variables so that a dictionary is
created at the point of testing using the lists.