Created two functions to edit a startup item's file path
Created helper function edit_startup_path to edit
the file path for a specific startup item and
return it. Updated edit_startup_item to call the
function and also to update the startup_item
dictionary as well as display the updated items.

Also created a helper function called choose_file
to display a file chooser dialog using Tkinter.
In edit_startup_path, the user is allowed to
choose the file chooser dialog or to manually type
in the full file path.