Added new option in main loop for creating new startup_data.json files
Modified main loop to allow user to create a new
startup file with default JSON data, or create a
new file with data that they specify. This rewrote
the max choices and choice numbers.

For the function json_creator, added a boolean
parameter to specify which type of new file to
create. Updated the docstring to explain the
changes. Also updated the call to create_json_data
to pass in the boolean parameter.

For the function create_json_data, added a second
boolean parameter to specify which type of new
file to create. Updated the docstring to explain
the changes. Also updated the if block to account
for the new scenario. The two boolean parameters
are now set to False by default. The first
parameter has to be changed to True to create a
new file. If that's the case, then the second
parameter is taken into account. If it's true, use
default data for the new JSON file. If it's false,
call a new function called get_startup_data that
will serve to get the data the user wants to add.