Added new option when editing a startup item
Added option to delete a startup item. Also,
refactored the if statement for that question so
there aren't so many OR statements.

For example, the if statement now checks to see if
the user_choice variable is an alpha character and
if so, if it's in the list ["Q", "D", "A"],
standing for Quit, Delete, and Add.