Created new helper function and function rename
Created a new helper function, new_file_chooser, so in the main loop, instead of 5 options, there's only 4 and only 1 for creating a new file. If the user selects 1, then this helper function is called and will ask the user whether to create a new startup file with default data or user-provided input.

Redid the main loop to reflect the logic change. Also renamed choose_file to edit_file_chooser to better reflect the function's purpose and to differentiate from new_file_chooser.