Added, edited and fixed code related to creating a new startup file
In the function new_file_chooser, corrected the
question because it was displaying incorrectly.

In the function generate_default_startup_data,
commented out all the code that uses the ENUM
classes to create JSON data and just copied the
constant DEFAULT_JSON. Did this because the code
using the ENUM classes only creates one startup
item but have now decided to use 3 startup items
as default.

In the function generate_user_startup_data, used
the ENUM classes to generate startup data with
total items as 0 and a blank items array. Also,
added docstring data.

Finally, made a minor edit in the function
generate_json_data.