Coded for edge case where total items in startup data is greater than 0 but the items array is blank
While testing out previous changes, stumbled on an
edge case where the items array is blank but the
total items key has a value of 1, or not 0. This
was most likely caused by an accidental manual
editing of the test_data.json file and should
never happen in production, but added code to
both the prettify_json and json_editor functions
to handle this edge case. Now, if there are no
actual startup items to display, both the main
loop choices to view and edit the existing data
will state there are no startup items to display.
This covers both the edge case and also a regular
test case where a user might choose to view or
edit startup data that doesn't exist.