Refactored user choice code in both json_editor and edit_startup_item
Refactored the code governing the loop and the
user choice of options to make them in line with
the main loop using the format:

Choose from one of the following:
[1] Choice 1
[2] Choice 2

What would you like to do?

In json_editor, added code to allow for the
variability number of startup items. For example,
if there are 3 startup items, the first 3 choices
will be "Edit startup item #" with the next 3
being "add a new item, delete an item, return to
previous menu". The last 3 will always be that but
will be numbered according to how many startup
items there are.