Finished coding helper function user_menu_chooser
Decided to have helper function just print out
the menu choices passed in, confirm if the choice
is valid, and if not, print out that a valid
choice is needed. The return integer will be 0 by
default and only changed on a valid choice.

All caller functions will be responsible for
looping through and checking the return integer
from user_menu_chooser.

Also, updated the main loop, the function
new_file_chooser to use a menu instead of a single
question sentence, the function json_editor, and
the function edit_startup_item. Made sure all
menus have an option to return to the previous
menu.