Updated edit_startup_arguments
Edited the docstring for the function
edit_startup_arguments as well as wrote in the
initial if-block. Currently, if there are
arguments passed in, the function will return
the arguments list as in by passing.

If there are no arguments, the user is asked if
they would like to add some arguments. The default
reply is no, but if the user enters Y for yes,
a new function called add_startup_arguments will
be called.