Created helper function add_startup_arguments
Started coding helper function
add_startup_arguments which can be called any time
a customer wants to add arguments to a startup
item. For now, the function prints out that the
functionality hasn't been implemented and returns
the passed in arguments list, which defaults to
blank just in case one isn't passed in.