Added new menu option in function edit_startup_item
Added an option to show the startup data for the
specific item being worked on. This will allow the
user to recall the startup item data without
needing to go back up one level.

Also, moved the location of the function
add_startup_arguments so it's before all the edit
functions when going from top to bottom.