Updated edit_startup_arguments
Updated menu loop for edit_startup_arguments to
work correctly when a user doesn't make a change.
Also added code to both update the returned
arguments list and the displayed menu when a user
does change an argument. Finally, added a menu
item to add a new argument. Need to code in the
logic for that.