Minor code and menu changes
Changed the menu in new_file_chooser to not have
the option to return to the previous menu. Changed
the menu in json_editor to say "return to the main
menu" so menus lower down can say "return to the
previous menu" and it won't be confusing.

Also refactored the code for the edit_startup
functions when they check if the user input is
blank or not. Instead of comparing the user input
to "", the if blocks now just check if the user
input variable has a value or not.