Updated and refactored edit_startup_arguments
Added the options to delete an argument and to
save changes. Refactored the code so it's more
streamlined and less confusing due to the fact
that the user now has at least 5 choices:

- edit an argument
- add an argument
- delete an argument
- save any changes
- cancel and return to the previous menu

The code before the refactor would have created
multiple nested if-blocks within the menu loop to
check for each possible option. Currently, the
code has 3 nested if-blocks within the menu loop
and 5 levels of nested flow statements from the
function level:

if > while > if > if > if